# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProposalForge & AgentLead API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /al/properties:
    get:
      operationId: listAlProperties
      tags: [al-properties]
      summary: List properties
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [available, under_contract, sold, rented, off_market]
        - in: query
          name: type
          schema:
            type: string
            enum: [sale, rent]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlProperty"
    post:
      operationId: createAlProperty
      tags: [al-properties]
      summary: Create property
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlPropertyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlProperty"
  /al/properties/{id}:
    get:
      operationId: getAlProperty
      tags: [al-properties]
      summary: Get property
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlProperty"
    put:
      operationId: updateAlProperty
      tags: [al-properties]
      summary: Update property
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlPropertyBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlProperty"
    delete:
      operationId: deleteAlProperty
      tags: [al-properties]
      summary: Delete property
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AL LEADS ─────────────────────────────────────────────────────────────────
  /al/leads:
    get:
      operationId: listAlLeads
      tags: [al-leads]
      summary: List leads
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, contacted, qualified, lost]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlLead"
    post:
      operationId: createAlLead
      tags: [al-leads]
      summary: Create lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlLeadBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlLead"
  /al/leads/{id}:
    get:
      operationId: getAlLead
      tags: [al-leads]
      summary: Get lead
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlLead"
    put:
      operationId: updateAlLead
      tags: [al-leads]
      summary: Update lead
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlLeadBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlLead"
    delete:
      operationId: deleteAlLead
      tags: [al-leads]
      summary: Delete lead
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AL DEALS ─────────────────────────────────────────────────────────────────
  /al/deals:
    get:
      operationId: listAlDeals
      tags: [al-deals]
      summary: List deals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: stage
          schema:
            type: string
            enum: [prospect, showing, offer, negotiation, closing, closed_won, closed_lost]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlDeal"
    post:
      operationId: createAlDeal
      tags: [al-deals]
      summary: Create deal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlDealBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlDeal"
  /al/deals/{id}:
    get:
      operationId: getAlDeal
      tags: [al-deals]
      summary: Get deal
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlDeal"
    put:
      operationId: updateAlDeal
      tags: [al-deals]
      summary: Update deal
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlDealBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlDeal"
    delete:
      operationId: deleteAlDeal
      tags: [al-deals]
      summary: Delete deal
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /al/deals/{id}/stage:
    patch:
      operationId: updateAlDealStage
      tags: [al-deals]
      summary: Update deal stage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlDeal"

  # ─── AL REMINDERS ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateStageBody:
      type: object
      properties:
        stage:
          type: string
      required: [stage]

    # ── ProposalForge ─────────────────────────────────────────────────────────
...
    AlProperty:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        locationText:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        price:
          type: number
        currency:
          type: string
        type:
          type: string
          enum: [sale, rent]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        sizeM2:
          type: number
        status:
          type: string
          enum: [available, under_contract, sold, rented, off_market]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, locationText, city, price, currency, type, status, createdAt, updatedAt]
    CreateAlPropertyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        locationText:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        price:
          type: number
        currency:
          type: string
        type:
          type: string
          enum: [sale, rent]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        sizeM2:
          type: number
        status:
          type: string
          enum: [available, under_contract, sold, rented, off_market]
        notes:
          type: string
      required: [title, locationText, city, price, currency, type, status]
    AlLead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, lost]
        budget:
          type: number
        currency:
          type: string
        interestedInType:
          type: string
          enum: [sale, rent, both]
        preferredCity:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, status, createdAt, updatedAt]
    CreateAlLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, lost]
        budget:
          type: number
        currency:
          type: string
        interestedInType:
          type: string
          enum: [sale, rent, both]
        preferredCity:
          type: string
        notes:
          type: string
      required: [name, status]
    AlDeal:
      type: object
      properties:
        id:
          type: string
        leadId:
          type: string
        leadName:
          type: string
        propertyId:
          type: string
        propertyTitle:
          type: string
        title:
          type: string
        stage:
          type: string
          enum: [prospect, showing, offer, negotiation, closing, closed_won, closed_lost]
        dealValue:
          type: number
        currency:
          type: string
        commissionRate:
          type: number
        commissionAmount:
          type: number
        expectedCloseDate:
          type: string
          format: date
        closedAt:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, stage, dealValue, currency, createdAt, updatedAt]
    CreateAlDealBody:
      type: object
      properties:
        leadId:
          type: string
        propertyId:
          type: string
        title:
          type: string
        stage:
          type: string
          enum: [prospect, showing, offer, negotiation, closing, closed_won, closed_lost]
        dealValue:
          type: number
        currency:
          type: string
        commissionRate:
          type: number
        expectedCloseDate:
          type: string
          format: date
        notes:
          type: string
      required: [title, stage, dealValue, currency]
...